requirements.txt
Django==5.1.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
whitenoise==6.7.0
gunicorn==22.0.0
djangorestframework
djangorestframework-simplejwt